Split base-iterators and chainables into separate files
