+++ lib/base-iterators/isiterable.js
