+++ lib/iterate.js
@@ -1,7 +1,7 @@
 const proxy = require("./proxy");
 const sequences = require("./sequences");
 const chainables = require("./chainables");
-const isIterable = require("./isiterable");
+const { isIterable } = require("./base-iterators");
 
 function defineSequence(name) {
 	return function(...args) {
